Separated functions from model_classification.py into three new files: ensemble_functions.py, preprocess_identifiers.py, and process_features.py. Re-configured tests and imports to work with the new path. Updated README.md to reflect new instructions for configuring the tagger and for testing that it works.
